requirements.txt
requests~=2.25.1
SQLAlchemy~=1.3.23
python-ldap~=3.3.1
